requirements.txt
ag-ui-protocol==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
argcomplete==3.6.2
attrs==25.4.0
boto3==1.40.50
botocore==1.40.50
brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cohere==5.18.0
colorama==0.4.6
ddgs==9.6.1
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
eval-type-backport==0.2.2
executing==2.2.1
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastavro==1.12.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
genai-prices==0.0.31
google-auth==2.41.1
google-genai==1.43.0
googleapis-common-protos==1.70.0
griffe==1.14.0
groq==0.32.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.0
invoke==2.2.1
jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
logfire==4.13.0
logfire-api==4.13.0
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.17.0
mdurl==0.1.2
mistralai==1.9.11
multidict==6.7.0
nexus-rpc==1.1.0
openai==2.3.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
primp==0.15.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.0
pydantic-ai==1.0.17
pydantic-ai-slim==1.0.17
pydantic-core==2.41.1
pydantic-evals==1.0.17
pydantic-graph==1.0.17
pydantic-settings==2.11.0
pygments==2.19.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
sentry-sdk==2.41.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sse-starlette==3.0.2
starlette==0.48.0
temporalio==1.18.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.2
types-protobuf==6.32.1.20250918
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
